+++ src/main/java/mission/post/application/PostController.java
@@ -2,10 +2,10 @@ package mission.post.application;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import mission.post.business.PostService;
 import mission.post.dto.CreatePostRequest;
 import mission.post.dto.PostCreateResponse;
 import mission.post.dto.UpdatePostRequest;
-import mission.post.business.PostService;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
